# backend/ms-geolocalisation/src/main/resources/openapi/api.yml
openapi: 3.0.3
info:
  title: PGVPT - MS Géolocalisation API
  description: Propriétaire des coordonnées, du référentiel administratif, des zones géographiques et touristiques, des points
    d'accès et des recherches spatiales. Une géolocalisation cible exactement un patrimoine ou une entreprise; les références
    sont des UUID.
  version: 14.0.0
  contact:
    name: PGVPT
    email: contact@pgvpt.sn
servers:
  - url: http://localhost:3002/api/v1/geolocalisations
    description: Environnement local
  - url: https://api.pgvpt.sn/geolocalisation/api/v1
    description: Environnement de production
...
paths:
...
  /patrimoines/{patrimoineId}/points-acces:
    get:
      tags:
        - Geospatial
      summary: Lister les points d'accès d'un patrimoine
      operationId: getPointsAcces
      responses:
        '200':
          description: Points d'accès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PointAcces'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Geospatial
      summary: Ajouter un point d'accès
      operationId: createPointAcces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointAccesCreate'
      responses:
        '201':
          description: Point d'accès créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointAcces'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    parameters:
      - $ref: '#/components/parameters/PatrimoineId'
  /points-acces/{id}:
    put:
      tags:
        - Geospatial
      summary: Modifier un point d'accès
      operationId: updatePointAcces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointAccesCreate'
      responses:
        '200':
          description: Point d'accès modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointAcces'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Geospatial
      summary: Supprimer un point d'accès
      operationId: deletePointAcces
      responses:
        '204':
          description: Point d'accès supprimé
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    parameters:
      - $ref: '#/components/parameters/PointAccesId'
  /geolocalisations/recherche/rayon:
    get:
      tags:
        - Geospatial
      summary: Rechercher les patrimoines dans un rayon
      operationId: rechercherDansRayon
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: rayonKm
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: 0
            exclusiveMinimum: true
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TypePatrimoine'
        - name: categorie
          in: query
          schema:
            $ref: '#/components/schemas/CategoriePatrimoine'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Patrimoines trouvés dans le rayon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePatrimoineGeographique'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /geolocalisations/recherche/bbox:
    get:
      tags:
        - Geospatial
      summary: Rechercher les patrimoines dans une emprise géographique
      operationId: rechercherDansBoundingBox
      parameters:
        - name: minLatitude
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: minLongitude
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: maxLatitude
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: maxLongitude
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Patrimoines situés dans l'emprise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePatrimoineGeographique'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /geolocalisations/distance:
    get:
      tags:
        - Geospatial
      summary: Calculer la distance entre deux positions
      operationId: calculerDistance
      parameters:
        - name: latitudeDepart
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: longitudeDepart
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: latitudeArrivee
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: longitudeArrivee
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Distance calculée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Distance'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /geolocalisations/distance/patrimoines:
    get:
      tags:
        - Geospatial
      summary: Calculer la distance entre deux patrimoines
      operationId: calculerDistancePatrimoines
      parameters:
        - name: patrimoineIdDepart
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: patrimoineIdArrivee
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Distance entre les patrimoines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Distance'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cartographie/patrimoines:
    get:
      tags:
        - Geospatial
      summary: Obtenir les patrimoines pour affichage cartographique
      operationId: getPatrimoinesCarte
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TypePatrimoine'
        - name: categorie
          in: query
          schema:
            $ref: '#/components/schemas/CategoriePatrimoine'
      responses:
        '200':
          description: Données cartographiques
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeaturePatrimoine'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cartographie/geojson:
    get:
      tags:
        - Geospatial
      summary: Exporter les patrimoines au format GeoJSON
      operationId: exportGeoJson
      responses:
        '200':
          description: Collection GeoJSON des patrimoines
          content:
            application/geo+json:
              schema:
                $ref: '#/components/schemas/GeoJsonFeatureCollection'
        '500':
          $ref: '#/components/responses/InternalServerError'
      parameters: []
...
  /geolocalisations/recherche:
    get:
      tags:
        - Geospatial
      summary: Rechercher les ressources géolocalisées selon plusieurs critères
      operationId: searchGeolocalisations
      parameters:
        - name: latitude
          in: query
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: longitude
          in: query
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: rayonKm
          in: query
          schema:
            type: number
            format: double
            minimum: 0
            exclusiveMinimum: true
        - name: minLatitude
          in: query
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: maxLatitude
          in: query
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: minLongitude
          in: query
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: maxLongitude
          in: query
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: localite
          in: query
          schema:
            type: string
        - name: motCle
          in: query
          schema:
            type: array
            items:
              type: string
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TypePatrimoine'
        - name: categorie
          in: query
          schema:
            $ref: '#/components/schemas/CategoriePatrimoine'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Résultats de recherche
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePatrimoineGeographique'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
...
    PatrimoineId:
      name: patrimoineId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PointAccesId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      description: Numéro de page, indexé à partir de 0.
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: Nombre maximal d’éléments par page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Unauthorized:
      description: Authentification requise
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Accès interdit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: Erreur interne
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnprocessableEntity:
      description: Données métier invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    BaseEntity:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      description: Classe racine commune de toutes les entités persistantes du PGVPT.
...
    PatrimoineGeographique:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
      properties:
        nom:
          type: string
        type:
          $ref: '#/components/schemas/TypePatrimoine'
        categorie:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        distanceKm:
          type: number
          format: double
        zoneTouristique:
          $ref: '#/components/schemas/ZoneTouristiqueReference'
        village:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/VillageReference'
        quartier:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/QuartierReference'
        patrimoineId:
          type: string
          format: uuid
          nullable: true
        entrepriseId:
          type: string
          format: uuid
          nullable: true
    Distance:
      type: object
      properties:
        distanceKm:
          type: number
          format: double
          minimum: 0
          example: 12.45
        distanceMetres:
          type: number
          format: double
          minimum: 0
          example: 12450
        dureeEstimeeMinutes:
          type: integer
          minimum: 0
        modeTransport:
          $ref: '#/components/schemas/ModeTransport'
        origine:
          $ref: '#/components/schemas/Coordonnee'
        destination:
          $ref: '#/components/schemas/Coordonnee'
      required:
        - distanceKm
        - distanceMetres
        - origine
        - destination
      description: Résultat calculé de distance entre deux positions.
    Coordonnee:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
      description: Coordonnée géographique en WGS84. Latitude comprise entre -90 et 90 et longitude entre -180 et 180.
    PointAcces:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
      properties:
        patrimoineId:
          type: string
          format: uuid
        nom:
          type: string
        type:
          $ref: '#/components/schemas/TypePointAcces'
        description:
          type: string
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        adresse:
          type: string
        distancePatrimoineMetres:
          type: number
          format: double
          minimum: 0
        accessiblePMR:
          type: boolean
        parkingDisponible:
          type: boolean
        transportPublic:
          type: boolean
        horaires:
          type: string
    PointAccesCreate:
      type: object
      required:
        - nom
        - type
        - latitude
        - longitude
      properties:
        nom:
          type: string
        type:
          $ref: '#/components/schemas/TypePointAcces'
        description:
          type: string
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        adresse:
          type: string
        accessiblePMR:
          type: boolean
        parkingDisponible:
          type: boolean
        transportPublic:
          type: boolean
...
    FeaturePatrimoine:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
      properties:
        patrimoineId:
          type: string
          format: uuid
        nom:
          type: string
        type:
          $ref: '#/components/schemas/TypePatrimoine'
        categorie:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        icone:
          type: string
          format: uri
        couleur:
          type: string
        description:
          type: string
    GeoJsonFeatureCollection:
      type: object
      required:
        - type
        - features
      properties:
        type:
          type: string
          enum:
            - FeatureCollection
        features:
          type: array
          items:
            $ref: '#/components/schemas/GeoJsonFeature'
    GeoJsonFeature:
      type: object
      required:
        - type
        - geometry
        - properties
      properties:
        type:
          type: string
          enum:
            - Feature
        geometry:
          $ref: '#/components/schemas/GeoJsonPoint'
        properties:
          type: object
          additionalProperties: true
    GeoJsonPoint:
      type: object
      required:
        - type
        - coordinates
      properties:
        type:
          type: string
          enum:
            - Point
        coordinates:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
          description: Coordonnées GeoJSON dans l’ordre [longitude, latitude].
      description: Géométrie GeoJSON de type Point. Les coordonnées sont exprimées dans l’ordre [longitude, latitude].
    TypePointAcces:
      type: string
      enum:
        - ENTREE_PRINCIPALE
        - ENTREE_SECONDAIRE
        - PARKING
        - QUAI
        - EMBARCADERE
        - GARE
        - ARRET_BUS
        - AEROPORT
        - PORT
        - CENTRE_VISITEUR
        - AUTRE
...
    ModeTransport:
      type: string
      enum:
        - A_PIED
        - VOITURE
        - TAXI
        - BUS
        - TRAIN
        - BATEAU
        - VELO
        - MOTO
        - AUTRE
...
    ApiError:
      type: object
      required:
        - timestamp
        - status
        - code
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        path:
          type: string
        details:
          type: object
          additionalProperties: true
...
    VillageReference:
      type: object
      required:
        - id
        - code
        - nom
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: VIL001
        nom:
          type: string
          example: Île de Gorée
    QuartierReference:
      type: object
      required:
        - id
        - code
        - nom
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: Q001
        nom:
          type: string
          example: Centre
...
    ZoneTouristiqueReference:
      type: object
      required:
        - id
        - code
        - nom
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: ZT-CASAMANCE
        nom:
          type: string
          example: Zone touristique de Casamance
...
    TypePatrimoine:
      type: string
      enum:
        - SITE_NATUREL
        - MUSEE
        - MONUMENT
      description: Type de patrimoine géré par ms-patrimoine.
    CategoriePatrimoine:
      type: string
      enum:
        - PATRIMOINE_NATUREL
        - PATRIMOINE_HISTORIQUE
        - PATRIMOINE_CULTUREL
        - PATRIMOINE_ARCHITECTURAL
        - PATRIMOINE_RELIGIEUX
        - PATRIMOINE_COLONIAL
        - PATRIMOINE_ARCHEOLOGIQUE
        - PATRIMOINE_MUSEAL
        - PATRIMOINE_MEMORIEL
        - PATRIMOINE_MARITIME
        - PATRIMOINE_RURAL
        - AUTRE
...
    PagePatrimoineGeographique:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PatrimoineGeographique'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []